+++ frontends/android/BitBoxApp/app/src/main/java/ch/shiftcrypto/bitboxapp/Util.java
@@ -3,6 +3,7 @@ package ch.shiftcrypto.bitboxapp;
 import android.app.Application;
 import android.content.Intent;
 import android.net.Uri;
+import android.util.Log;
 
 public class Util {
     public static void systemOpen(Application application, String url) throws Exception {
@@ -12,5 +13,9 @@ public class Util {
         if (intent.resolveActivity(application.getPackageManager()) != null) {
             application.startActivity(intent);
         }
+    }
+
+    public static void log(String msg) {
+        Log.d("ch.shiftcrypto.bitboxapp", msg);
     }
 }